Add files via upload
This repository features a comprehensive collection of data structures and algorithms implemented in C. Focused on efficiency and clarity, it serves as a valuable resource for mastering foundational concepts in algorithms and data structures. Curated by a 2nd-year Computer Science and Engineering student, it provides clean, optimized code examples for both academic and practical applications.